+++ migrations/migrate-requeue-count.js
@@ -1,0 +1,1 @@
+db.deltaFile.updateMany({requeueCount: null}, {$set: {"requeueCount": NumberInt("0")}})